memory-store/services/api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
psycopg2-binary==2.9.9
pgvector==0.2.4
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.0

